Fix Empty Package Check
